middleware folder, client side cache moved from cache.py to a file in middleware
