+++ src/app/core/cache.py
@@ -6,12 +6,8 @@ import re
 from fastapi import Request, Response
 from fastapi.encoders import jsonable_encoder
 from redis.asyncio import Redis, ConnectionPool
-from fastapi import FastAPI
-from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
 
 from app.core.exceptions import CacheIdentificationInferenceError, InvalidRequestError
-
-# --------------- server side caching ---------------
 
 pool: ConnectionPool | None = None
 client: Redis | None = None
@@ -334,59 +330,3 @@ def cache(
         return inner
 
     return wrapper
-
-# --------------- client side caching ---------------
-
-class ClientCacheMiddleware(BaseHTTPMiddleware):
-    """
-    Middleware to set the `Cache-Control` header for client-side caching on all responses.
-    
-    Parameters
-    ----------
-    app: FastAPI 
-        The FastAPI application instance.
-    max_age: int, optional 
-        Duration (in seconds) for which the response should be cached. Defaults to 60 seconds.
-
-    Attributes
-    ----------
-    max_age: int
-        Duration (in seconds) for which the response should be cached.
-
-    Methods
-    -------
-    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
-        Process the request and set the `Cache-Control` header in the response.
-
-    Note
-    ----
-        - The `Cache-Control` header instructs clients (e.g., browsers) to cache the response for the specified duration.
-    """
-    
-    def __init__(self, app: FastAPI, max_age: int = 60) -> None:
-        super().__init__(app)
-        self.max_age = max_age
-
-    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
-        """
-        Process the request and set the `Cache-Control` header in the response.
-        
-        Parameters
-        ----------
-        request: Request
-            The incoming request.
-        call_next: RequestResponseEndpoint
-            The next middleware or route handler in the processing chain.
-            
-        Returns
-        -------
-        Response
-            The response object with the `Cache-Control` header set.
-
-        Note
-        ----
-            - This method is automatically called by Starlette for processing the request-response cycle.
-        """
-        response: Response = await call_next(request)
-        response.headers['Cache-Control'] = f"public, max-age={self.max_age}"
-        return response
